refactor(db): extract checkpoint follower protocol
